pyproject.toml
[project]
name = "crypto-key"
version = "0.1.0"
description = ""
authors = [
    {name = "adrlksv",email = "Qwerty123Andrey@bk.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "bcrypt (==4.0.1)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "cryptography (>=44.0.3,<45.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
